requirements.txt
qdrant-edge-py>=0.6.0
fastembed>=0.4.0
openai-whisper>=20231117
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.30.0
sounddevice>=0.4.6
soundfile>=0.12.0
scipy>=1.11.0
torch>=2.0.0
pydub>=0.25.1
